auth-ms/pyproject.toml
[tool.poetry]
name = "auth-ms"
version = "0.1.0"
description = ""
authors = ["Peskov Sergey <peskovdev@proton.me>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
python-dotenv = "^1.0.0"
pyjwt = "^2.6.0"
flask-cors = "^3.0.10"
flask-sqlalchemy = "^3.0.3"
psycopg2 = "^2.9.5"
flask-migrate = "^4.0.4"


[tool.poetry.group.dev.dependencies]
pynvim = "^0.4.3"
isort = "^5.12.0"
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
